unit test regions

+++ src/controllers/regionsController.js
@@ -1,5 +1,3 @@
-const Region = require("../models/region");
-
 function regionHATEOAS(region) {
     const { code, name, country } = region;
     return {
@@ -23,7 +21,7 @@ function regionHATEOAS(region) {
     }
 }
 
-async function all(req, res, next) {
+async function all(req, res, next, Region) {
     try {
         let regions = await Region.findAll();
         regions = regions.map(regionHATEOAS);
@@ -33,7 +31,7 @@ async function all(req, res, next) {
     }
 }
 
-async function allByCountry(req, res, next) {
+async function allByCountry(req, res, next, Region) {
     try {
         let regions = await Region.findAll({
             where: {
@@ -51,7 +49,7 @@ async function allByCountry(req, res, next) {
     }
 }
 
-async function getRegion(req, res, next) {
+async function getRegion(req, res, next, Region) {
     try {
         let region = await Region.findByPk(req.params.region);
         if(region) {
@@ -65,40 +63,54 @@ async function getRegion(req, res, next) {
     }
 }
 
-async function deleteRegion(req, res, next) {
+async function deleteRegion(req, res, next, Region, City) {
     try {
-        let deleted = await Region.destroy({
-            where: {
-                code: req.params.region
+        const { country, region } = req.params;
+        let cities = await City.findOne({
+            where: { country, region }
+        })
+        if(!cities) {
+            let deleted = await Region.destroy({
+                where: {
+                    code: req.params.region
+                }
+            });
+            if(deleted) {
+                res.status(204).send(`deleted region with code ${req.params.region}`);
+            } else {
+                res.status(404).send(`No region found with code ${req.params.region}`);
             }
-        });
-        if(deleted) {
-            res.status(204).send(`deleted region with code ${req.params.region}`);
         } else {
-            res.status(404).send(`No region found with code ${req.params.region}`);
+            res.status(405).send("Cannot delete region with associated cities");
         }
     } catch(err) {
         next(new Error(`Error deleting region: ${err.message}`));
     }
 }
 
-async function createRegion(req,res,next) {
+async function createRegion(req, res, next, Region, Country) {
     try {
-        let newRegion = {
-            code: req.body.code,
-            name: req.body.name,
-            country: req.params.country
+        let country = await Country.findByPk(req.params.country);
+        if(country) {
+            let newRegion = {
+                code: req.body.code,
+                name: req.body.name,
+                country: req.params.country
+            }
+            let region = await Region.create(newRegion);
+            region = regionHATEOAS(region);
+            res.status(201).send(region);
+        } else {
+            res.status(405).send("Cannot create region for non existent country");
         }
-        let region = await Region.create(newRegion);
-        region = regionHATEOAS(region);
-        res.status(200).send(region);
     } catch(err) {
         next(new Error(`Error creating region: ${err.message}`));
     }
 }
 
-async function updateRegion(req, res, next) {
+async function updateRegion(req, res, next, Region) {
     try {
+        let status;
         let region = {
             name: req.body.name,
             country: req.body.country
@@ -109,12 +121,18 @@ async function updateRegion(req, res, next) {
             }
         })
         if(updatedRows > 0) {
-            let region = await Region.findByPk(req.params.region);
-            region = regionHATEOAS(region);
-            res.status(200).send(region);
+            region = await Region.findByPk(req.params.region);
+            status = 200;
         } else {
-            res.status(404).send(`No region found with code ${req.params.region}`);
+            let newRegion = {
+                code: req.params.region,
+                ...region
+            }
+            region = await Region.create(newRegion);
+            status = 201;
         }
+        region = regionHATEOAS(region);
+        res.status(status).send(region);
     } catch(err) {
         next(new Error(`Error updating or creating region ${err.message}`));
     }
